% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_color_scales.R
\docType{data}
\name{c25_colors}
\alias{c25_colors}
\title{25 vizuálisan jól elkülöníthető szín kódja}
\format{
Karaktervektor, 25 elem.
}
\usage{
c25_colors
}
\description{
Ez a vektor 25 színt tartalmaz, amelyek jól megkülönböztethetők
grafikonokon, kategóriák színezéséhez stb.
}
\details{
A színek részben R beépített névvel (\code{"dodgerblue2"}, \code{"gold1"}, stb.),
részben hex kóddal vannak megadva.
}
\examples{
# Példa: oszlopdiagram színezése
barplot(1:25, col = c25_colors)

}
\keyword{datasets}
