% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_based_on_TSP.R
\name{order_based_on_TSP}
\alias{order_based_on_TSP}
\title{Sorrendezés TSP alapján}
\usage{
order_based_on_TSP(distance_mx)
}
\arguments{
\item{distance_mx}{Szimmetruikus négyzetes numerikus mátrix. A kisebb értékek közeli
elemeket jelentenek. Ha vannak sornevek, a függvény ezek sorrendjét adja
vissza; ha nincsenek, automatikus neveket használ.}
}
\value{
Karaktervektor: az elemek TSP-alapú sorrendje.
}
\description{
Távolságmátrix alapján keres egy kis költségű Hamilton-utat a
travelling salesman problem (TSP) átalakításával. Hasznos például
korrelációs mátrixok heatmap ábrázolásánál a sorok és oszlopok
sorrendjének meghatározásához.
}
\details{
A függvény egy dummy csomópontot ad a távolságmátrixhoz, TSP körutat
számol, majd a dummy csomópont eltávolításával Hamilton-utat nyer vissza.
}
\examples{
set.seed(1)
x <- matrix(stats::rnorm(25), nrow = 5)
cor_mx <- stats::cor(x)
dist_mx <- as.matrix(stats::dist(cor_mx))
rownames(dist_mx) <- paste0("v", seq_len(nrow(dist_mx)))
colnames(dist_mx) <- rownames(dist_mx)

order_based_on_TSP(dist_mx)

}
