% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_handling.R
\name{outlier_threshold_at_normal_sample}
\alias{outlier_threshold_at_normal_sample}
\title{Abból kiindulva, hogy normális eloszású a minta, kiszámolja hogy ennyi
mintaelemszám esetén hol lenne aa  olyan kétoldali határ, amelyen kívül egy
N-elemű normális mintában csak ritkán várnánk értéket.}
\usage{
outlier_threshold_at_normal_sample(x, conf = 0.95)
}
\description{
Az átlag és a szórás becslésére robosztus becslést használ, hogy az outlierek ne
zavarjanak be
}
\details{
@param x numeric vektor.
@param conf numeric vektor.
@return thr 2 elemű vektorban egy intervallummal. Az első érték kell a kisebb

Példa:
dat |>  mutate(burden=winsorize(burden, outlier_threshold_at_normal_sample(burden,0.99)))
}
