% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_helpers.R
\name{print_as_string_list}
\alias{print_as_string_list}
\title{Karaktervektor idézőjelezése és vesszővel elválasztott formázása}
\usage{
print_as_string_list(x)
}
\arguments{
\item{x}{karaktervektor.}
}
\value{
Látható kimenet a konzolra (\code{cat}), visszatérési értéke \code{NULL}.
}
\description{
A bemeneti karaktervektor elemeit idézőjelek közé teszi és
vesszővel elválasztva kiírja a konzolra (copy–paste barát).
}
\examples{
# print_as_string_list(c("alma", "korte", "szilva"))
}
